+++ README.md
@@ -52,7 +52,7 @@ If you need to remove absolutely everything
 docker compose down -v --rmi all \
     && docker system prune \
     && docker compose build --no-cache \
-    && docker compose up --force-recreate
+    && docker compose up --force-recreate -d
 ```
 
 If you would like to run docker in a deteched state, add the `-d` flag like this
