+++ interactive/shared/configuration_utils.py
@@ -5,9 +5,7 @@ import os
 import pandas as pd
 import shutil
 import fjson
-from api import models
 from zipfile import ZipFile
-
 
 def get_session_path(session_id):
     '''Returns the absolute path to the configuration folder for a given session id'''
@@ -45,14 +43,6 @@ def get_zip_folder_path(session_id):
     return path
 
 
-def get_partmc_zip_folder_path(session_id):
-    '''Returns the folder for zipped partmc for a given session'''
-    os.makedirs(os.environ['PARTMC_ZIP_DIR'], exist_ok=True)
-    path = os.path.join(os.environ['PARTMC_ZIP_DIR'], session_id)
-    os.makedirs(path, exist_ok=True)
-    return path
-
-
 def get_unzip_folder_path(session_id):
     '''Returns the folder to unzip configurations into'''
     os.makedirs(os.environ['MUSIC_BOX_CONFIG_DIR'], exist_ok=True)
@@ -66,19 +56,9 @@ def get_zip_file_path(session_id):
     return os.path.join(get_zip_folder_path(session_id), "config.zip")
 
 
-def get_partmc_zip_file_path(session_id):
-    '''Returns the path for a zip file for a given session id'''
-    return os.path.join(get_partmc_zip_folder_path(session_id), "config.zip")
-
-
 def remove_zip_folder(session_id):
     '''Removes the folder holding a compressed configuration for a session'''
     shutil.rmtree(get_zip_folder_path(session_id))
-
-
-def remove_zip_folder_partmc(session_id):
-    '''Removes the folder holding a compressed partmc for a session'''
-    shutil.rmtree(get_partmc_zip_folder_path(session_id))
 
 
 def replace_large_ints(data):
@@ -251,15 +231,6 @@ def compress_configuration(session_id):
     make_archive(config_folder, zip_file_path)
     remove_session_folder(session_id)
 
-
-def compress_partmc(session_id):
-    '''Creates a compressed file holding the partmc output from a given session id'''
-    model = models.ModelRun.objects.get(uid=session_id)
-    partmc_folder = model.results['partmc_output_path']
-    partmc_zip_file_path = get_partmc_zip_file_path(session_id)
-    logging.info(
-        f'Compressing configuration: {partmc_folder} to: {partmc_zip_file_path}')
-    make_archive(partmc_folder, partmc_zip_file_path)
 
 
 def make_archive(source, destination):
