+++ pyproject.toml
@@ -22,6 +22,7 @@ pyOpenSSL = "^22.0"
 python = "^3.9"
 scipy = "^1.7"
 setuptools = "^70.0"
+Werkzeug = "^2.0"
 
 pypartmc = { version = "^1.3.3", optional = true }
 acom_music_box = { version = "^2.2.1", optional = true }
